backend/requirements.txt
Flask
Flask-Cors
PyMySQL
Flask-SQLAlchemy
python-dotenv
PyJWT